Finish switching to Go from Python
There is only one `index.csv` now, which is the source of truth for
everything. `tags.csv` and `tag_mapper.csv` serve for making it easier
for me to input data into `index.csv` directly, so that I don't have
to remember the exact name for every tag.

Based on the `index.csv`, the Go script renders an HTML template that
does pretty much the exact same thing as before, conceptually, but
it's much simpler now.

The dependency on Python is no more, finally. All I have to do to
generate the website now is:

    go run .

And that's it!